requirements.txt
flask~=1.1.2
flask-restx~=0.5.1
Werkzeug~=1.0.1
bcrypt~=4.1.2
PyJWT~=2.8.0